requirements.txt
Django>=2.2,<3.0
django-environ==0.4.5
django-filter==2.1.0
djangorestframework==3.9.4
ipython
psycopg2-binary==2.8.3
